ml-service/requirements.txt
# KRISHI-AI ML Service Dependencies

# Web Framework
<<<<<<< HEAD
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Data Processing
pyyaml>=6.0.1
python-Levenshtein>=0.23.0

# OCR
pytesseract>=0.3.10
Pillow>=10.2.0

# ML (optional - for enhanced ranking)
scikit-learn>=1.4.0
numpy>=1.26.3
joblib>=1.3.2

# HTTP Client
httpx>=0.26.0

# Database
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Code Quality
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
=======
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
pyyaml==6.0.1
python-Levenshtein==0.23.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# ML (optional - for enhanced ranking)
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2

# HTTP Client
httpx==0.26.0

# Database
psycopg2-binary==2.9.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
>>>>>>> f00a66f0fc28ee1666e2d7573bc8ca1e3b8da83f
